+++ app/flask_blueprints/admin_blueprints/manage_users.py
@@ -1,11 +1,17 @@
 from flask import Blueprint, render_template, request, redirect, url_for, session, flash
-from app import mysql
-from flask_blueprints.admin_blueprint import admin_bp
+from flask_blueprints.admin_blueprint import admin_bp, get_db_connection
 
 
 @admin_bp.route('/pasazerowie')
 def admin_pasazerowie():
-    cursor = mysql.connection.cursor()
+    if 'role' not in session or session['role'] != 'admin':
+        flash('Brak dostępu')
+        return redirect(url_for('admin.login'))
+    
+
+    conn = get_db_connection('admin')
+    cursor = conn.cursor()
+
     cursor.execute("SELECT id_pasażera, imie, nazwisko, mail, telefon, haslo FROM pasazerowie")
     users = cursor.fetchall()
     cursor.close()
@@ -15,7 +21,13 @@ def admin_pasazerowie():
 
 @admin_bp.route('/pasazerowie/<int:user_id>/bilety')
 def pokaz_bilety_pasazera(user_id):
-    cursor = mysql.connection.cursor()
+    if 'role' not in session or session['role'] != 'admin':
+        flash('Brak dostępu')
+        return redirect(url_for('admin.login'))
+    
+    conn = get_db_connection('admin')
+    cursor = conn.cursor()
+
     cursor.execute("SELECT id_biletu, id_pasażera, id_połączenia, cena, ulgi FROM bilety WHERE id_pasażera = %s", (user_id,))
     tickets = cursor.fetchall()
 
@@ -28,10 +40,16 @@ def pokaz_bilety_pasazera(user_id):
 
 
 @admin_bp.route('/pasazerowie/<int:user_id>/bilety/<int:ticket_id>/usun', methods=['POST'])
-def usun_bilet(user_id, ticket_id):   
-    cursor = mysql.connection.cursor()
+def usun_bilet(user_id, ticket_id):  
+    if 'role' not in session or session['role'] != 'admin':
+        flash('Brak dostępu')
+        return redirect(url_for('admin.login'))
+    
+    conn = get_db_connection('admin')
+    cursor = conn.cursor()
+    
     cursor.execute("DELETE FROM bilety WHERE id_biletu = %s", (ticket_id,))
-    mysql.connection.commit()
+    conn.commit()
     cursor.close()
     return redirect(url_for('admin.pokaz_bilety_pasazera', user_id=user_id))
 
@@ -39,14 +57,22 @@ def usun_bilet(user_id, ticket_id):
 
 @admin_bp.route('/pasazerowie/<int:user_id>/bilety/<int:ticket_id>/edytuj', methods=['POST'])
 def edytuj_bilet(user_id, ticket_id):
+    if 'role' not in session or session['role'] != 'admin':
+        flash('Brak dostępu')
+        return redirect(url_for('admin.login'))
+    
     cena = request.form.get('cena')
     ulga = request.form.get('ulga') 
 
-    cursor = mysql.connection.cursor()
+    conn = get_db_connection('admin')
+    cursor = conn.cursor()
+
     cursor.execute("""
         UPDATE bilety SET cena = %s, ulgi = %s WHERE id_biletu = %s
     """, (cena, ulga, ticket_id))
-    mysql.connection.commit()
+    
+    conn.commit()
+    
     cursor.close()
 
     return redirect(url_for('admin.pokaz_bilety_pasazera', user_id=user_id))
@@ -54,8 +80,13 @@ def edytuj_bilet(user_id, ticket_id):
 
 
 @admin_bp.route('/pasazerowie/<int:user_id>', methods=['GET', 'POST'])
-def edytuj_pasazera(user_id):   
-    cursor = mysql.connection.cursor()
+def edytuj_pasazera(user_id):
+    if 'role' not in session or session['role'] != 'admin':
+        flash('Brak dostępu')
+        return redirect(url_for('admin.login'))   
+    
+    conn = get_db_connection('admin')
+    cursor = conn.cursor()
 
     cursor.execute("""SELECT * FROM pasazerowie WHERE id_pasażera=%s""", (user_id,))
     user = cursor.fetchall()
@@ -74,8 +105,8 @@ def edytuj_pasazera(user_id):
                 telefon = %s
             WHERE id_pasażera = %s
         """, (imie, naziwsko, email, telefon, user_id))
-        
-        mysql.connection.commit()
+        conn.commit()
+
         cursor.close()
         return redirect(url_for('admin.admin_pasazerowie'))
 
@@ -90,9 +121,15 @@ def edytuj_pasazera(user_id):
 
 @admin_bp.route('/pasazerowie/<int:user_id>/usun', methods=['POST'])
 def usun_pasazera(user_id):   
-    cursor = mysql.connection.cursor()
+    if 'role' not in session or session['role'] != 'admin':
+        flash('Brak dostępu')
+        return redirect(url_for('admin.login'))
+    
+    conn = get_db_connection('admin')
+    cursor = conn.cursor()
+    
     cursor.execute("DELETE FROM pasazerowie WHERE id_pasażera = %s", (user_id,))
-    mysql.connection.commit()
+    conn.commit()
     cursor.close()
     return redirect(url_for('admin.admin_pasazerowie', user_id=user_id))
